requirements.txt
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1 